feat: slip execution endpoint
